server/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.8.0
django-environ==0.12.0
django-password-validators==1.7.3
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
shapely==2.0.6
